Homework

+++ src/main/java/com/javaguru/student_Vladislav_Pipirin/lesson_2/homework/level_3/Task_11.java
@@ -12,7 +12,7 @@ public class Task_11 {
         int firstNumber = input.nextInt();
         int secondNumber = input.nextInt();
         int thirdNumber = input.nextInt();
-        double averageResult = (double) (firstNumber + secondNumber + thirdNumber) / 3;
+        double averageResult = (firstNumber + secondNumber + thirdNumber) / 3.0;
         System.out.println(" average =  " + averageResult);
 
     }
